third_party/WebKit/LayoutTests/fast/dom/shadow/host-pseudo-class.html

 <!doctype html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/shadow-dom.js"></script>
<style>
:host {
   background-color: red;
}
</style>
</head>
<body class='mytheme'>
  <div>
    <div id='sandbox'></div>
  </div>
  <pre id='console'></pre>
<script>
description('Test whether :host matches a shadow host correctly.');

var sandbox = document.getElementById('sandbox');

function backgroundColorOf(selector) {
    return window.getComputedStyle(getNodeInTreeOfTrees(selector)).backgroundColor;
}

function backgroundColorShouldBe(selector, expected) {
    shouldBeEqualToString('backgroundColorOf(\'' + selector + '\')', expected);
}

function cleanUp() {
    sandbox.innerHTML = '';
}

debug(':host out of shadow tree should not match any shadow hosts.');

sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('div', {},
                document.createTextNode('Hello')))));

backgroundColorShouldBe('host', 'rgba(0, 0, 0, 0)');

cleanUp();

debug(':host with * should not match any shadow hosts.');

sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode('*:host { background-color: green; }')),
            createDOM('div', {},
                document.createTextNode('Hello')))));

backgroundColorShouldBe('host', 'rgb(0, 128, 0)');

cleanUp();

debug(':host with tag selector should not match any shadow hosts.');

sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode('div:host { background-color: green; }')),
            createDOM('div', {},
                document.createTextNode('Hello')))));

backgroundColorShouldBe('host', 'rgba(0, 0, 0, 0)');

cleanUp();

debug(':host with class selector should not match any shadow hosts.');

sandbox.appendChild(
    createDOM('div', {'id': 'host', 'class': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode('.host:host { background-color: green; }')),
            createDOM('div', {},
                document.createTextNode('Hello')))));

backgroundColorShouldBe('host', 'rgba(0, 0, 0, 0)');

cleanUp();

debug(':host with id selector should not match any shadow hosts.');

sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode('#host:host { background-color: green; }')),
            createDOM('div', {},
                document.createTextNode('Hello')))));

backgroundColorShouldBe('host', 'rgba(0, 0, 0, 0)');

cleanUp();

debug(':host with attribute selector should not match any shadow hosts.');

sandbox.appendChild(
    createDOM('div', {'id': 'host', 'foo': 'bar'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode('[foo=bar]:host { background-color: green; }')),
            createDOM('div', {},
                document.createTextNode('Hello')))));

backgroundColorShouldBe('host', 'rgba(0, 0, 0, 0)');

cleanUp();

debug(':host in a shadow tree should match its shadow host.');

sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode(':host(body.mytheme) { background-color: green; }')),
            createDOM('div', {},
                document.createTextNode('Hello')))));

backgroundColorShouldBe('host', 'rgb(0, 128, 0)');

cleanUp();

debug(':host takes simple selectors and matches when one of the simple selectors matches.');

sandbox.appendChild(
    createDOM('div', {'id': 'parentOfHost'},
        createDOM('div', {'id': 'host'},
            createShadowRoot(
                createDOM('style', {},
                    document.createTextNode(':host(body:not(.mytheme), span, div#parentOfHost) { background-color: green; }')),
                createDOM('div', {},
                    document.createTextNode('Hello'))))));

backgroundColorShouldBe('host', 'rgb(0, 128, 0)');

cleanUp();

debug(':host matches a shadow host in just a nested shadow tree, not all enclosing shadow trees.');

sandbox.appendChild(
    createDOM('div', {'id': 'host1'},
        createShadowRoot(
            createDOM('div', {'id': 'host2'},
                createShadowRoot(
                    createDOM('style', {},
                        document.createTextNode(':host { background-color: green; }')),
                    createDOM('div', {},
                        document.createTextNode('Hello')))))));

backgroundColorShouldBe('host1', 'rgba(0, 0, 0, 0)');
backgroundColorShouldBe('host1/host2', 'rgb(0, 128, 0)');

debug(':host matches based on a composed tree.');

sandbox.appendChild(
    createDOM('div', {'id': 'parentOfHost'},
        createShadowRoot(
            createDOM('span', {'id': 'spanA'},
                createDOM('content', {}))),
        createDOM('div', {'id': 'host'},
            createShadowRoot(
                createDOM('style', {},
                    document.createTextNode(':host(span#spanA) > div { background-color: green; }')),
                createDOM('div', {'id': 'target'},
                    document.createTextNode('Hello'))))));

backgroundColorShouldBe('host/target', 'rgb(0, 128, 0)');

cleanUp();

debug(':host matches based on a composed tree when having multiple shadow roots.');

sandbox.appendChild(
    createDOM('div', {'id': 'parentOfHost'},
        createShadowRoot(
            createDOM('span', {'id': 'spanA'},
                document.createTextNode('no content, no shadow'))),
        createShadowRoot(
            createDOM('span', {'id': 'spanB'},
                createDOM('content', {}))),
        createShadowRoot(
            createDOM('span', {'id': 'spanC'},
                createDOM('shadow', {}))),
        createDOM('div', {'id': 'host'},
            createShadowRoot(
                createDOM('style', {},
                    document.createTextNode(':host(span#spanA) > #targetA { background-color: red; }')),
                createDOM('style', {},
                    document.createTextNode(':host(span#spanB) > #targetB { background-color: green; }')),
                createDOM('style', {},
                    document.createTextNode(':host(span#spanC) > #targetC { background-color: green; }')),
                createDOM('div', {'id': 'targetA'},
                    document.createTextNode('Hello')),
                createDOM('div', {'id': 'targetB'},
                    document.createTextNode('Hello')),
                createDOM('div', {'id': 'targetC'},
                    document.createTextNode('Hello'))))));

backgroundColorShouldBe('host/targetA', 'rgba(0, 0, 0, 0)');
backgroundColorShouldBe('host/targetB', 'rgb(0, 128, 0)');
backgroundColorShouldBe('host/targetC', 'rgb(0, 128, 0)');

cleanUp();

debug(':host is updated when its matched ancestor changes className or id.');

sandbox.appendChild(
    createDOM('div', {'id': 'parentOfHost', 'class': 'sometheme' },
        createDOM('div', {'id': 'host'},
            createShadowRoot(
                createDOM('style', {},
                    document.createTextNode(':host(div#parentOfHost.sometheme) { background-color: green; }')),
                createDOM('div', {},
                    document.createTextNode('Hello'))))));

backgroundColorShouldBe('host', 'rgb(0, 128, 0)');
document.getElementById('parentOfHost').className = 'mytheme';
backgroundColorShouldBe('host', 'rgba(0, 0, 0, 0)');

cleanUp();

sandbox.appendChild(
    createDOM('div', {'id': 'parentOfHost', 'class': 'sometheme' },
        createDOM('div', {'id': 'host'},
            createShadowRoot(
                createDOM('style', {},
                    document.createTextNode(':host(div#parentOfHost.sometheme) { background-color: green; }')),
                createDOM('div', {},
                    document.createTextNode('Hello'))))));

backgroundColorShouldBe('host', 'rgb(0, 128, 0)');
var parentOfHost = document.getElementById('parentOfHost');
parentOfHost.id = 'differntIdValue';
backgroundColorShouldBe('host', 'rgba(0, 0, 0, 0)');

cleanUp();

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
